requirements.txt
face-recognition>=1.3.0
opencv-python>=4.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.0
pytest>=7.4.0
hypothesis>=6.92.0
Pillow>=10.0.0
